🚚 [RUMF-848] move types and constants to respect the new ESLint rule
