+++ packages/rum-recorder/src/constants.ts
@@ -1,0 +1,6 @@
+export const PRIVACY_ATTR_NAME = 'data-dd-privacy'
+export const PRIVACY_ATTR_VALUE_HIDDEN = 'hidden'
+export const PRIVACY_ATTR_VALUE_INPUT_IGNORED = 'input-ignored'
+
+export const PRIVACY_CLASS_HIDDEN = 'dd-privacy-hidden'
+export const PRIVACY_CLASS_INPUT_IGNORED = 'dd-privacy-input-ignored'
