+++ packages/rum-recorder/src/domain/privacy.ts
@@ -1,9 +1,10 @@
-export const PRIVACY_ATTR_NAME = 'data-dd-privacy'
-export const PRIVACY_ATTR_VALUE_HIDDEN = 'hidden'
-export const PRIVACY_ATTR_VALUE_INPUT_IGNORED = 'input-ignored'
-
-export const PRIVACY_CLASS_HIDDEN = 'dd-privacy-hidden'
-export const PRIVACY_CLASS_INPUT_IGNORED = 'dd-privacy-input-ignored'
+import {
+  PRIVACY_ATTR_NAME,
+  PRIVACY_ATTR_VALUE_HIDDEN,
+  PRIVACY_CLASS_HIDDEN,
+  PRIVACY_ATTR_VALUE_INPUT_IGNORED,
+  PRIVACY_CLASS_INPUT_IGNORED,
+} from '../constants'
 
 // PRIVACY_INPUT_TYPES_TO_IGNORE defines the input types whose input
 // events we want to ignore by default, as they often contain PII.
