+++ packages/rum-recorder/src/internal.ts
@@ -22,6 +22,6 @@ export {
   PRIVACY_ATTR_VALUE_INPUT_IGNORED,
   PRIVACY_CLASS_HIDDEN,
   PRIVACY_CLASS_INPUT_IGNORED,
-} from './domain/privacy'
+} from './constants'
 
 export * from './types'
